ml-service/app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
psycopg2-binary
pydantic
numpy
